Fix ci

+++ trackers/bytetrack/byte_tracker.py
@@ -18,7 +18,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score, cls):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=np.float32)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
